fixed access key verify, updated PutUserData and made GetUserData

+++ lockr/src/service/data.cpp
@@ -9,7 +9,14 @@
 namespace lockr {
     static int maxData = 1024*1024; // 1MB
 
-    bool RemoveAllData(const std::string& userId, const std::string& companyId) {
+    bool RemoveAllData(const std::string& userId, const std::string& companyId, std::string& outError) {
+        if(!DB::Exists("user", bsoncxx::builder::basic::make_document(
+                bsoncxx::builder::basic::kvp("_id", userId)
+        ))) {
+            outError = "User could not be found";
+            return false;
+        }
+
         return DB::DeleteOne("data", bsoncxx::builder::basic::make_document(
                   bsoncxx::builder::basic::kvp("user_id", userId),
                   bsoncxx::builder::basic::kvp("company_id", companyId)
@@ -17,16 +24,17 @@ namespace lockr {
     }
 
     //TODO
-    bool GetData(const std::string& userId, nlohmann::json& outJson, const std::string& companyId, const std::string fields[]) {
-        return true;
+    bool GetData(const std::string& userId, const std::string& companyId, const std::string fields[], std::string& outError, nlohmann::json& outJson) {
+
     }
 
-    bool GetAllData(const std::string& userId, nlohmann::json& outJson, const std::string& companyId) {
+    bool GetAllData(const std::string& userId, const std::string& companyId, std::string& outError, nlohmann::json& outJson) {
         auto doc = DB::getOne("data", bsoncxx::builder::basic::make_document(
                   bsoncxx::builder::basic::kvp("user_id", userId),
                   bsoncxx::builder::basic::kvp("company_id", companyId)
                   ));
         if (!doc) {
+            outError = "User could not be found";
             outJson = nullptr;
             return false;
         }
@@ -36,9 +44,19 @@ namespace lockr {
         return true;
     }
 
-    bool ReplaceData(const std::string& userId, const nlohmann::json& json, const std::string& companyId) {
+    bool ReplaceData(const std::string& userId, const nlohmann::json& json, const std::string& companyId, std::string& outError) {
+        if(!DB::Exists("user", bsoncxx::builder::basic::make_document(
+                bsoncxx::builder::basic::kvp("_id", userId)
+        ))) {
+            outError = "User could not be found";
+            return false;
+        }
+
         const bsoncxx::document::value jsonDoc = bsoncxx::from_json(json.dump());
-        if (jsonDoc.view().length() > maxData) return false;
+        if (jsonDoc.view().length() > maxData){
+            outError = "User's data limit has been reached";
+            return false;
+        };
 
         if (!DB::Exists("data", bsoncxx::builder::basic::make_document(
             bsoncxx::builder::basic::kvp("user_id", userId),
@@ -60,14 +78,14 @@ namespace lockr {
         return true;
     }
 
-    bool MergeData(const std::string& userId, const nlohmann::json& json, const std::string& companyId) {
+    bool MergeData(const std::string& userId, const nlohmann::json& json, const std::string& companyId, std::string& outError) {
         nlohmann::json currentJson;
-        GetAllData(userId, currentJson, companyId);
+        if(!GetAllData(userId, companyId, outError, currentJson)) return false;
         if (currentJson == nullptr) {
-            return ReplaceData(userId, json, companyId);
+            return ReplaceData(userId, json, companyId, outError);
         }
         currentJson.merge_patch(json);
 
-        return ReplaceData(userId, currentJson, companyId);
+        return ReplaceData(userId, currentJson, companyId, outError);
     }
 }
